+++ src/MoBi.Presentation/Views/ICloneBuildingBlocksToModuleView.cs
@@ -1,0 +1,11 @@
+﻿using MoBi.Presentation.DTO;
+using MoBi.Presentation.Presenter;
+using OSPSuite.Presentation.Views;
+
+namespace MoBi.Presentation.Views
+{
+   public interface ICloneBuildingBlocksToModuleView : IModalView<ICloneBuildingBlocksToModulePresenter>
+   {
+      void BindTo(CloneBuildingBlocksToModuleDTO cloneBuildingBlocksToModuleDTO);
+   }
+}
